@@ -1,11 +1,21 @@
 module lib.kevinlin.info/proton
 
-go 1.16
+go 1.17
 
 require (
 	github.com/gorilla/websocket v1.4.2
 	github.com/stretchr/testify v1.7.0
 	golang.org/x/lint v0.0.0-20210508222113-6edffad5e616
-	golang.org/x/tools v0.1.3
-	lib.kevinlin.info/aperture v0.0.0-20210508182034-84a61db2c3ec
+	golang.org/x/tools v0.1.7
+	lib.kevinlin.info/aperture v0.0.0-20211003163055-bbeab5f73620
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/mod v0.4.2 // indirect
+	golang.org/x/net v0.0.0-20210929193557-e81a3d93ecf6 // indirect
+	golang.org/x/sys v0.0.0-20211003122950-b1ebd4e1001c // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
