@@ -1,11 +1,19 @@
 module wg
 
-go 1.16
+go 1.17
 
 require (
-	github.com/buger/goterm v1.0.0
+	github.com/buger/goterm v1.0.3
 	golang.org/x/lint v0.0.0-20210508222113-6edffad5e616
-	golang.org/x/net v0.0.0-20210614182718-04defd469f4e
-	golang.org/x/tools v0.1.3
-	lib.kevinlin.info/proton v0.0.0-20210618200354-62b58235add2
+	golang.org/x/net v0.0.0-20210929193557-e81a3d93ecf6
+	golang.org/x/tools v0.1.7
+	lib.kevinlin.info/proton v0.0.0-20211003164559-948e3d69a11e
+)
+
+require (
+	github.com/gorilla/websocket v1.4.2 // indirect
+	golang.org/x/mod v0.4.2 // indirect
+	golang.org/x/sys v0.0.0-20211003122950-b1ebd4e1001c // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	lib.kevinlin.info/aperture v0.0.0-20211003163055-bbeab5f73620 // indirect
 )
