@@ -1,9 +1,12 @@
 package webgrep
 
 import (
+	"encoding/base64"
 	"net/http"
+	"strings"
 
 	"lib.kevinlin.info/proton"
+	"lib.kevinlin.info/proton/supercharged"
 )
 
 // Client is a webgrep API client; effectively, a single layer of abstraction above a Supercharged
@@ -22,7 +25,7 @@ func NewClient(cfg *proton.Config) (*Client, error) {
 	return &Client{proton.NewClient(parsed)}, nil
 }
 
-// Search executes a search query.
+// Search executes a code search query.
 func (c *Client) Search(request *SearchQueryRequest) (*SearchQueryResponse, error) {
 	var resp SearchQueryResponse
 
@@ -31,6 +34,27 @@ func (c *Client) Search(request *SearchQueryRequest) (*SearchQueryResponse, erro
 	}
 
 	return &resp, nil
+}
+
+// Source executes a source code payload query.
+func (c *Client) Source(request *SourceQueryRequest) (*SourceQueryResponse, error) {
+	var resp string
+
+	if err := c.sc.DoHTTP(http.MethodGet, EndpointSource, request, &resp); err != nil {
+		return nil, err
+	}
+
+	decoded, err := base64.StdEncoding.DecodeString(resp)
+	if err != nil {
+		return nil, supercharged.WrapError(err)
+	}
+
+	return &SourceQueryResponse{
+		Repository: request.Repository,
+		Version:    request.Version,
+		Path:       request.Path,
+		Lines:      strings.Split(string(decoded), "\n"),
+	}, nil
 }
 
 // Metadata requests metadata about the webgrep instance.
